api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.4
cryptography==37.0.4
fastapi==0.77.1
greenlet==2.0.1
h11==0.13.0
idna==3.3
isodate==0.6.1
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
pathspec==0.9.0
platformdirs==2.5.2
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pydub==0.25.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.44
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.11
uvicorn==0.16.0
watchgod==0.7
youtube-dl==2021.12.17
zipp==3.6.0
